Refactor x/bank according to ADR 031
