+++ buf.yaml
@@ -14,6 +14,7 @@ lint:
     - COMMENT_FIELD
     - SERVICE_SUFFIX
     - PACKAGE_VERSION_SUFFIX
+    - RPC_REQUEST_STANDARD_NAME
   ignore:
     - tendermint
     - gogoproto
