+++ proto/cosmos/bank/v1beta1/tx.proto
@@ -7,6 +7,15 @@ import "cosmos/bank/v1beta1/bank.proto";
 
 option go_package = "github.com/cosmos/cosmos-sdk/x/bank/types";
 
+// Msg defines the bank Msg service.
+service Msg {
+  // Send defines a method for sending coins from one account to another account.
+  rpc Send(MsgSend) returns (MsgSendResponse);
+
+  // MultiSend defines a method for sending coins from some accounts to other accounts.
+  rpc MultiSend(MsgMultiSend) returns (MsgMultiSendResponse);
+}
+
 // MsgSend represents a message to send coins from one account to another.
 message MsgSend {
   option (gogoproto.equal)           = false;
@@ -18,6 +27,9 @@ message MsgSend {
       [(gogoproto.nullable) = false, (gogoproto.castrepeated) = "github.com/cosmos/cosmos-sdk/types.Coins"];
 }
 
+// MsgSendResponse defines the Msg/Send response type.
+message MsgSendResponse { }
+
 // MsgMultiSend represents an arbitrary multi-in, multi-out send message.
 message MsgMultiSend {
   option (gogoproto.equal) = false;
@@ -25,3 +37,6 @@ message MsgMultiSend {
   repeated Input inputs   = 1 [(gogoproto.nullable) = false];
   repeated Output outputs = 2 [(gogoproto.nullable) = false];
 }
+
+// MsgMultiSendResponse defines the Msg/MultiSend response type.
+message MsgMultiSendResponse { }
